[IMP] spreadsheet: convert cell popover store to owl plugin
Convert the cell popover store to owl plugin.

The conversion is not the greatest yet, as we need to access the
plugin in the helpers that only take the env as argument, which
we cannot do ATM.

Task: 6306326
