+++ src/components/grid/grid.ts
@@ -77,9 +77,9 @@ import { useWheelHandler } from "../helpers/wheel_hook";
 import { ZoomedMouseEvent } from "../helpers/zoom";
 import { Highlight } from "../highlight/highlight/highlight";
 import { MenuPopover, MenuState } from "../menu_popover/menu_popover";
+import { CellPopoverPlugin } from "../owl_plugins/cell_popover_plugin";
 import { DelayedHoveredCellPlugin } from "../owl_plugins/delayed_hovered_cell_plugin";
 import { PaintFormatStore } from "../paint_format_button/paint_format_store";
-import { CellPopoverStore } from "../popover/cell_popover_store";
 import { Popover } from "../popover/popover";
 import { types } from "../props_validation";
 import { HorizontalScrollBar } from "../scrollbar/scrollbar_horizontal";
@@ -147,7 +147,7 @@ export class Grid extends Component<SpreadsheetChildEnv> {
   private gridRef = signal<HTMLElement | null>(null);
   private canvasRef = signal<HTMLElement | null>(null);
   private highlightStore!: Store<HighlightStore>;
-  private cellPopovers!: Store<CellPopoverStore>;
+  private cellPopovers = plugin(CellPopoverPlugin);
   private composerFocusStore!: Store<ComposerFocusStore>;
   private DOMFocusableElementStore!: Store<DOMFocusableElementStore>;
   private paintFormatStore!: Store<PaintFormatStore>;
@@ -191,7 +191,6 @@ export class Grid extends Component<SpreadsheetChildEnv> {
       this.moveCanvas(deltaX, deltaY);
       this.hoveredCell.clear();
     });
-    this.cellPopovers = useStore(CellPopoverStore);
 
     useLayoutEffect(
       (isMainPanelOpen, isSecondaryPanelOpen) => {
