+++ src/components/popover/cell_popover_store.ts
@@ -1,101 +1,12 @@
 import { plugin } from "@odoo/owl";
-import { positionToZone } from "../../helpers/zones";
-import { cellPopoverRegistry } from "../../registries/cell_popovers_registry";
 import { SpreadsheetStore } from "../../stores/spreadsheet_store";
-import {
-  CellPopoverType,
-  ClosedCellPopover,
-  OpenCellPopover,
-  PositionedCellPopoverComponent,
-} from "../../types/cell_popovers";
-import { Command } from "../../types/commands";
-import { CellPosition, Position } from "../../types/misc";
-import { Rect } from "../../types/rendering";
-import { DelayedHoveredCellPlugin } from "../owl_plugins/delayed_hovered_cell_plugin";
+import { CellPopoverPlugin } from "../owl_plugins/cell_popover_plugin";
 
 export class CellPopoverStore extends SpreadsheetStore {
   mutators = ["open", "close"] as const;
 
-  private persistentPopover?: CellPosition & { type: CellPopoverType };
-
-  protected hoveredCell = plugin(DelayedHoveredCellPlugin);
-
-  handle(cmd: Command) {
-    switch (cmd.type) {
-      case "ACTIVATE_SHEET":
-        this.close();
-    }
-  }
-
-  open({ col, row }: Position, type: CellPopoverType): void {
-    const sheetId = this.getters.getActiveSheetId();
-    if (!cellPopoverRegistry.contains(type)) {
-      return;
-    }
-    this.persistentPopover = { col, row, sheetId, type };
-  }
-
-  close() {
-    if (!this.persistentPopover) {
-      return "noStateChange";
-    }
-    this.persistentPopover = undefined;
-    return;
-  }
-
-  get persistentCellPopover(): OpenCellPopover | ClosedCellPopover {
-    return (
-      (this.persistentPopover && { isOpen: true, ...this.persistentPopover }) || { isOpen: false }
-    );
-  }
-
-  get isOpen() {
-    return this.persistentPopover !== undefined;
-  }
-
-  get cellPopover(): ClosedCellPopover | PositionedCellPopoverComponent {
-    const sheetId = this.getters.getActiveSheetId();
-
-    if (this.persistentPopover && this.getters.isVisibleInViewport(this.persistentPopover)) {
-      const position = this.getters.getMainCellPosition(this.persistentPopover);
-      const popover = cellPopoverRegistry
-        .get(this.persistentPopover.type)
-        .onOpen?.(position, this.getters);
-      return !popover?.isOpen
-        ? { isOpen: false }
-        : {
-            ...popover,
-            anchorRect: this.computePopoverAnchorRect(this.persistentPopover),
-          };
-    }
-    const col = this.hoveredCell.col();
-    const row = this.hoveredCell.row();
-    if (
-      col === undefined ||
-      row === undefined ||
-      !this.getters.isVisibleInViewport({ sheetId, col, row })
-    ) {
-      return { isOpen: false };
-    }
-    const position = this.getters.getMainCellPosition({ sheetId, col, row });
-    const popover = cellPopoverRegistry
-      .getAll()
-      .map((matcher) => matcher.onHover?.(position, this.getters))
-      .find((popover) => popover?.isOpen);
-    return !popover?.isOpen
-      ? { isOpen: false }
-      : {
-          ...popover,
-          anchorRect: this.computePopoverAnchorRect(position),
-        };
-  }
-
-  private computePopoverAnchorRect({ col, row }: Position): Rect {
-    const sheetId = this.getters.getActiveSheetId();
-    const merge = this.getters.getMerge({ sheetId, col, row });
-    if (merge) {
-      return this.getters.getVisibleRect(merge);
-    }
-    return this.getters.getVisibleRect(positionToZone({ col, row }));
-  }
+  // FIXME: the store is there because at the moment the plugins are not available in the helpers that take the env as argument.
+  // It is a bit hacky, and wrong because the plugin won't be scoped correctly.
+  // Remove the store once we can access the plugins from the helpers.
+  popoverPlugin = plugin(CellPopoverPlugin);
 }
